Fix v2.3 CI (#217)

+++ core/src/banking_stage.rs
@@ -744,7 +744,7 @@ impl BankingStage {
             let (finished_work_sender, finished_work_receiver) = unbounded();
 
             // Spawn the worker threads
-            let mut worker_metrics = Vec::with_capacity(num_workers as usize);
+            let mut worker_metrics = Vec::with_capacity(num_workers);
             for (index, work_receiver) in work_receivers.into_iter().enumerate() {
                 let id = (index as u32)
                     .saturating_add(NUM_VOTE_PROCESSING_THREADS)
